move shadcn use-mobile hook to hooks/common
