+++ front/components.json
@@ -15,7 +15,7 @@
     "utils": "@/lib/utils",
     "ui": "@/components/atoms/shadcn",
     "lib": "@/lib",
-    "hooks": "@/hooks"
+    "hooks": "@/hooks/common"
   },
   "iconLibrary": "lucide"
 }
