+++ front/src/components/atoms/shadcn/sidebar.tsx
@@ -5,7 +5,7 @@ import { Slot } from "@radix-ui/react-slot"
 import { VariantProps, cva } from "class-variance-authority"
 import { PanelLeftIcon } from "lucide-react"
 
-import { useIsMobile } from "@/hooks/use-mobile"
+import { useIsMobile } from "@/hooks/common/use-mobile"
 import { cn } from "@/lib/utils"
 import { Button } from "@/components/atoms/shadcn/button"
 import { Input } from "@/components/atoms/shadcn/input"
